+++ src/modules/groups.ts
@@ -1,0 +1,28 @@
+const groups = [
+  {
+    value: 1,
+    color: 'rgb(4, 255, 0)',
+  },
+  {
+    value: 2,
+    color: 'rgb(183, 228, 0)',
+  },
+  {
+    value: 3,
+    color: 'rgb(228, 213, 0)',
+  },
+  {
+    value: 4,
+    color: 'rgb(228, 129, 0)',
+  },
+  {
+    value: 5,
+    color: 'rgb(228, 53, 0)',
+  },
+  {
+    value: 6,
+    color: 'rgb(249, 0, 0)',
+  },
+];
+
+export default groups;
